# common: optimization pImpl data structures