--- a/src/loaders/external_webp/tvgWebpLoader.cpp
+++ b/src/loaders/external_webp/tvgWebpLoader.cpp
@@ -30,7 +30,6 @@
 /* Internal Class Implementation                                        */
 /************************************************************************/
 
-
 void WebpLoader::run(unsigned tid)
 {
     //TODO: acquire the current colorspace format & pre-multiplied alpha image.