--- a/src/renderer/gl_engine/tvgGlRenderTask.cpp
+++ b/src/renderer/gl_engine/tvgGlRenderTask.cpp
@@ -24,9 +24,6 @@
 #include "tvgGlProgram.h"
 #include "tvgGlRenderPass.h"
 
-/************************************************************************/
-/* External Class Implementation                                        */
-/************************************************************************/
 
 GlRenderTask::GlRenderTask(GlProgram* program, GlRenderTask* other): mProgram(program)
 {