--- a/src/renderer/gl_engine/tvgGlRenderer.cpp
+++ b/src/renderer/gl_engine/tvgGlRenderer.cpp
@@ -803,11 +803,9 @@ void GlRenderer::endRenderPass(RenderCompositor* cmp)
 }
 
 
-
 /************************************************************************/
 /* External Class Implementation                                        */
 /************************************************************************/
-
 
 bool GlRenderer::clear()
 {