--- a/src/renderer/tvgIteratorAccessor.h
+++ b/src/renderer/tvgIteratorAccessor.h
@@ -34,7 +34,7 @@ public:
     //Utility Method: Iterator Accessor
     static Iterator* iterator(const Paint* paint)
     {
-        return paint->pImpl->iterator();
+        return PAINT(paint)->iterator();
     }
 };