ThorVG pImpl idiom caused internal data to be scattered across hierarchical classes. This refactoring consolidates the data by inheriting pImpl internally, reducing memory allocation counts and eliminating unnecessary strategy methods.