@hermet 2024-12-11
local test:
- binary size (wasm): +1.1kb (+0.12%) increased
- performance(fps): 2ms(+4%) improved

```
before:
  Paint::unref(): 0.393s
  GI: 0.454s
  fps: 50ms
after:
  Paint::unref(): 0.388s
  GI: 0.33s
  fps: 48ms
```