@tinyjin 2024-12-11
local test with perf-test in web

**Test Condition:**
- 50 visible animations
- testing on localhost

**[Before]**
LightHouse Performance : 87
- Software : 48(43-53)FPS, 84(84-85)MB
- WebGPU : 38(35-42)FPS, 79(69-90)MB

**[After]**
LightHouse Performance : 84
- Software : 61(51-71)FPS, 87(80-94)MB
- WebGPU : 53(34-72)FPS, 84(79-90)MB

**Difference (After vs Before)**

Software
- FPS: `+27.1%` (48 → 61 FPS)
- Memory: `+2.96%` (84.5 → 87 MB)

WebGPU
- FPS: `+37.7%` (38.5 → 53 FPS)
- Memory: `+6.3%` (79.5 → 84.5 MB)

**Seed:** 
```
ZGFuY2luZ19zdGFyLGludGVsaWFfbG9nb19hbmltYXRpb24scnVmbyxyYWRhcixoYXBweV90cmlvLGhhbWJ1cmdlcixzd2luZ2luZyxzYWRfZW1vamksZHJpdmluZyw1MzE3LWZpcmV3b3JrZHMsbWVzc2FnZSxjYXRfbG9hZGVyLHN0YXJzdHJpcHMscmlwcGxlX2xvYWRpbmdfYW5pbWF0aW9uLHBvbHlzdGFyX2FuaW0sbWFza2luZywyNzc0Ni1qb3lwaXhlbHMtcGFydHlpbmctZmFjZS1lbW9qaS1hbmltYXRpb24sZ3VpdGFyLGZvb2RyYXRpbmcsc3dpbmdpbmcsd2F2ZXMsc25haWwsZWlkX211YmFyYWssdGV4dDIscnVmbyxyYWRhcix3b21hbixza3VsbGJveSx3YXRlcl9maWxsaW5nLGhvbGEsdWtfZmxhZyxtYXRlcmlhbF93YXZlX2xvYWRpbmcsc3Ryb2tlX2Rhc2gsd2F2ZXMsbGFfY29tbXVuYXV0ZSxnYW1lX2ZpbmlzaGVkLG1hc2tpbmcsbGlrZV9idXR0b24scnVmbyxmdW5fYW5pbWF0aW9uLGFic3RyYWN0X2NpcmNsZSxtYXRlcmlhbF93YXZlX2xvYWRpbmcsbWFza2luZyxjb25mZXR0aSxkYW5jaW5nX2Jvb2ssY2F0X2xvYWRlcixmb2NhbF90ZXN0LHJhZGFyLGR1Y2ssMTE1NTU=
```