@tinyjin 2024-12-12
I conducted performance test in web (perf-test)

**Test Condition:**
- 50 visible animations

**Before (Main)**
- Software: Avg. 44.5 FPS (40~49)
- WebGPU: Avg. 39.5 FPS (37~42)

**After (hermet/optimization not rebase)**
- Software: Avg. 47 FPS (39~55)
- WebGPU: Avg. 42.5 FPS (37~48)

**Percentage Change (After vs Before)**
- Software : +5.4% (44.5 → 47 FPS)
- WebGPU : +7.1% (39.5 → 42.5 FPS)